requirements.txt
--extra-index-url=https://pypi.nvidia.com

##############################
# RAPIDS (GPU-accelerated)
##############################
cudf-cu12==24.12.0
cuml-cu12==24.12.0
rmm-cu12==24.12.0
cupy-cuda12x==13.0.0

##############################
# CPU dependencies
##############################
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
statsmodels==0.14.1   
scipy==1.14.1
matplotlib==3.8.2        
seaborn==0.13.2          
